@@ -1,9 +1,7 @@
-import AbstractEvent from "../AbstractEvent";
+import AbstractEvent from '../AbstractEvent';
 
-class BasicEvent extends AbstractEvent
-{
-	public clone():BasicEvent
-	{
+class BasicEvent extends AbstractEvent {
+	public clone(): BasicEvent {
 		return new BasicEvent(this.type, this.bubbles, this.cancelable);
 	}
 }
