@@ -1,17 +1,17 @@
 // require all test files
-const testsContext = require.context(
+const testsContext = (<any> require).context(
 	'./',
 	true,
-	/Spec\.ts/
+	/Spec\.ts/,
 );
 
 testsContext.keys().forEach(testsContext);
 
 // require all source files
-const sourcesContext = require.context(
+const sourcesContext = (<any> require).context(
 	'../src/',
 	true,
-	/\.ts$/
+	/\.ts$/,
 );
 
 sourcesContext.keys().forEach(sourcesContext);
